Added solution - LeetHub

+++ Find Nth root of M - GFG/find-nth-root-of-m.cpp
@@ -10,7 +10,7 @@ class Solution{
 	    int low=1,high=m;
 	    
 	    while(low<=high){
-	        long long mid=low+(high-low)/2;
+	        int mid=low+(high-low)/2;
 	        if(pow(mid,n)==m)
 	        return mid;
 	        else if(pow(mid,n)<m)
